requirements.txt
langchain
langchain_core
langchain_community
langchain_groq
python-dotenv
opencv-python
mediapipe
numpy
Flask
pvporcupine
pvrecorder
